move Logo component to common components
